Add files via upload
New stuff!
kobolds (both regular and troopers) now drop random weapons, from wooden spears to mirthril swords or even the omnious sounding Soul Sucker: drinker of life!
added a tower to the city (as of yet incomplete as there is no prize at the top. Or a top. Climbing it is a hard challenge with no reward but it self)